+++ lib/pages/settings/settings_card.dart
@@ -1,6 +1,10 @@
+import 'package:beamer/beamer.dart';
 import 'package:flutter/material.dart';
+import 'package:lotti/database/database.dart';
+import 'package:lotti/get_it.dart';
 import 'package:lotti/services/nav_service.dart';
 import 'package:lotti/themes/theme.dart';
+import 'package:lotti/utils/consts.dart';
 
 class SettingsCard extends StatelessWidget {
   const SettingsCard({
@@ -16,8 +20,17 @@ class SettingsCard extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    void onTap() {
-      navigateNamedRoute(path);
+    void beamToNamed(String path) => context.beamToNamed(path);
+
+    Future<void> onTap() async {
+      final beamerNav =
+          await getIt<JournalDb>().getConfigFlag(enableBeamerNavFlag);
+
+      if (beamerNav) {
+        beamToNamed(path);
+      } else {
+        navigateNamedRoute(path);
+      }
     }
 
     return Card(