+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added:
-- New navigation (in progress, available via config flag)
+- New navigation using beamer (in progress, available via config flag)
+- Navigation in Settings > Tags using beamer
 
 ## [0.8.129] - 2022-08-17
 ### Changed: